+++ src/components/todo/todoList/TodoList.component.tsx
@@ -1,37 +1,20 @@
-import React, { useState, useEffect, useContext } from 'react'
-import { Spin } from 'antd';
-
-import DBContext from '../../../context/db.context'
-import { ITodoRecive } from '../../../interfaces'
+import React from 'react'
 
 import styles from './todoList.module.scss'
 import TodoListItem from '../todoListItem/TodoListItem.component'
+import { ITodoRecive } from '../../../interfaces'
 
 interface Props {
-  match: {
-    params: {
-      listId: string
-    }
+  todos: ITodoRecive,
+  list: {
+    title: string
   }
 }
 
-const TodoList = ({ match }: Props) => {
-  const db = useContext(DBContext)
-  const [todos, setTodos] = useState<ITodoRecive>([])
-
-  useEffect(() => {
-    db.get('todos')((collection: any) => collection.where(
-        'listId', '==', match.params.listId
-    )).then(setTodos)
-  }, [db, match.params.listId])
-
-  const list = db.lists?.find((list) => list.id === match.params.listId)
-
-  if (!list) return <Spin />
-
+const TodoList = ({ todos, list }: Props) => {
   return (
     <div className={styles.todoList}>
-      <p>{list?.title}</p>
+      <p>{list.title}</p>
       <ul>
         {
           todos?.map((todo) =>
